Extract State Machine from Game Loop
